--- a/.travis.yml
+++ b/.travis.yml
@@ -6,7 +6,7 @@ python:
 services: postgresql
 
 addons:
-  postgresql: "9.4"
+  postgresql: "9.6"
 
 env:
   global: