--- a/usaspending_api/awards/v2/filters/transaction.py
+++ b/usaspending_api/awards/v2/filters/transaction.py
@@ -5,6 +5,7 @@ from usaspending_api.awards.models import LegalEntity
 from usaspending_api.references.models import NAICS, PSC
 from usaspending_api.common.exceptions import InvalidParameterException
 from usaspending_api.awards.v2.filters.location_filter_geocode import geocode_filter_locations
+from usaspending_api.awards.v2.filters.filter_helpers import get_total_transaction_columns
 
 import logging
 logger = logging.getLogger(__name__)
@@ -249,37 +250,37 @@ def transaction_filter(filters):
 
         # award_amounts
         elif key == "award_amounts":
+            total_transaction_columns = get_total_transaction_columns(filters, TransactionNormalized)
             or_queryset = None
             queryset_init = False
             for v in value:
-                if v.get("lower_bound") is not None and v.get("upper_bound") is not None:
-                    if queryset_init:
-                        or_queryset |= TransactionNormalized.objects.filter(
-                            award__total_obligation__gt=v["lower_bound"],
-                            award__total_obligation__lt=v["upper_bound"]
-                        )
+                for column in total_transaction_columns:
+                    if v.get("lower_bound") is not None and v.get("upper_bound") is not None:
+                        bounds_dict = {
+                            '{}__gte'.format(column): v['lower_bound'],
+                            '{}__lte'.format(column): v['upper_bound']
+                        }
+                        if queryset_init:
+                            or_queryset |= TransactionNormalized.objects.filter(**bounds_dict)
+                        else:
+                            queryset_init = True
+                            or_queryset = TransactionNormalized.objects.filter(**bounds_dict)
+                    elif v.get("lower_bound") is not None:
+                        bounds_dict = {'{}__gte'.format(column): v['lower_bound']}
+                        if queryset_init:
+                            or_queryset |= TransactionNormalized.objects.filter(**bounds_dict)
+                        else:
+                            queryset_init = True
+                            or_queryset = TransactionNormalized.objects.filter(**bounds_dict)
+                    elif v.get("upper_bound") is not None:
+                        bounds_dict = {'{}__lte'.format(column): v['upper_bound']}
+                        if queryset_init:
+                            or_queryset |= TransactionNormalized.objects.filter(**bounds_dict)
+                        else:
+                            queryset_init = True
+                            or_queryset = TransactionNormalized.objects.filter(**bounds_dict)
                     else:
-                        queryset_init = True
-                        or_queryset = TransactionNormalized.objects.filter(award__total_obligation__gt=v["lower_bound"],
-                                                                           award__total_obligation__lt=v["upper_bound"])
-                elif v.get("lower_bound") is not None:
-                    if queryset_init:
-                        or_queryset |= TransactionNormalized.objects.filter(
-                            award__total_obligation__gt=v["lower_bound"]
-                        )
-                    else:
-                        queryset_init = True
-                        or_queryset = TransactionNormalized.objects.filter(award__total_obligation__gt=v["lower_bound"])
-                elif v.get("upper_bound") is not None:
-                    if queryset_init:
-                        or_queryset |= TransactionNormalized.objects.filter(
-                            award__total_obligation__lt=v["upper_bound"]
-                        )
-                    else:
-                        queryset_init = True
-                        or_queryset = TransactionNormalized.objects.filter(award__total_obligation__lt=v["upper_bound"])
-                else:
-                    raise InvalidParameterException('Invalid filter: award amount has incorrect object.')
+                        raise InvalidParameterException('Invalid filter: award amount has incorrect object.')
             if queryset_init:
                 queryset &= or_queryset