--- a/usaspending_api/download/v2/download_column_historical_lookups.py
+++ b/usaspending_api/download/v2/download_column_historical_lookups.py
@@ -369,7 +369,7 @@ query_paths = {
             'funding_sub_agency_name': 'latest_transaction__assistance_data__funding_sub_tier_agency_na',
             'last_modified_date': 'latest_transaction__assistance_data__modified_at',
             'non_federal_funding_amount': 'latest_transaction__assistance_data__non_federal_funding_amount',
-            'original_subsidy_cost': 'latest_transaction__assistance_data__original_loan_subsidy_cost',
+            'original_subsidy_cost': 'latest_transaction__original_loan_subsidy_cost',
             'period_of_performance_current_end_date': 'period_of_performance_current_end_date',
             'period_of_performance_start_date': 'period_of_performance_start_date',
             'primary_place_of_performance_city_name': 'latest_transaction__assistance_data__place_of_performance_city',
@@ -410,7 +410,8 @@ query_paths = {
             'recipient_zip_last_4_code': 'latest_transaction__assistance_data__legal_entity_zip_last4',
             'record_type_code': 'latest_transaction__assistance_data__record_type',
             'sai_number': 'latest_transaction__assistance_data__sai_number',
-            'total_funding_amount': 'latest_transaction__assistance_data__total_funding_amount'
+            'total_funding_amount': 'latest_transaction__assistance_data__total_funding_amount',
+            'total_subsidy_cost': 'total_subsidy_cost'
         }
     },
     'transaction': {
@@ -700,7 +701,7 @@ query_paths = {
             'last_modified_date': 'assistance_data__modified_at',
             'modification_number': 'modification_number',
             'non_federal_funding_amount': 'assistance_data__non_federal_funding_amount',
-            'original_subsidy_cost': 'assistance_data__original_loan_subsidy_cost',
+            'original_subsidy_cost': 'original_loan_subsidy_cost',
             'period_of_performance_current_end_date': 'period_of_performance_current_end_date',
             'period_of_performance_start_date': 'period_of_performance_start_date',
             'primary_place_of_performance_city_name': 'assistance_data__place_of_performance_city',
@@ -734,7 +735,8 @@ query_paths = {
             'recipient_zip_last_4_code': 'assistance_data__legal_entity_zip_last4',
             'record_type_code': 'assistance_data__record_type',
             'sai_number': 'assistance_data__sai_number',
-            'total_funding_amount': 'assistance_data__total_funding_amount'
+            'total_funding_amount': 'assistance_data__total_funding_amount',
+            'total_subsidy_cost': 'award__total_subsidy_cost'
         },
     },
     'subaward': {
@@ -1094,6 +1096,7 @@ human_names = {
             'total_funding_amount',
             'face_value_of_loan',
             'original_subsidy_cost',
+            'total_subsidy_cost',
             'period_of_performance_start_date',
             'period_of_performance_current_end_date',
             'awarding_agency_code',
@@ -1411,6 +1414,7 @@ human_names = {
             'total_funding_amount',
             'face_value_of_loan',
             'original_subsidy_cost',
+            'total_subsidy_cost',
             'action_date',
             'period_of_performance_start_date',
             'period_of_performance_current_end_date',