**High level description:**
Updating the award search endpoints to aggregate award amounts based on award type since loans should be aggregated on `original_loan_subsidy_cost`, and all other award types should be summed on `federal_action_obligation`. Updated the `award_amounts` filter to include loans.

When a specific "Award Type" filter is not selected, we will sum federal action obligation for everything that is NOT a loan, as"total obligation". We will then sum original loan subsidy cost for everything that IS a loan,  as "total subsidy cost". Finally, we will add "total obligation" and "total subsidy cost" to get the overall total to be displayed when there IS NO specific award type selected. The same applies when any type in addition to Loans is selected. If only Loans is selected, only original subsidy cost will be used. When loans is NOT selected AND any other type is selected, only federal action obligation will be used.

**Technical details:**
We've added case statements to award search endpoint queries that will apply the logic laid out above when aggregating totals. 

`sum_transaction_amount` function is used for the endpoints. This function is used to aggregate both transaction and award amounts. The award table already has `federal_action_obligation` and `original_loan_subsidy_cost` aggregated as `total_obligation` and `total_subsidy_cost`. Therefore we add an annotation for  transaction queries that creates these aggregated fields .

The award amount filter is also updated to specify the award amount based on the award type. The `get_total_transaction_columns` function  determines which field to put the upper and/or lower bound on based on whether the type is a loan or not.

**Link to JIRA Ticket:**
[DEV-489](https://federal-spending-transparency.atlassian.net/browse/DEV-489)
[DEV-490](https://federal-spending-transparency.atlassian.net/browse/DEV-490)

- [x] Backend review completed
- [x] Matview impact assessment completed
- [x] Frontend impact assessment completed
- [x] Data validation completed
```
Test 1 - Filter - Only Loans

Spending over time:
Expected : 
    2016 - 299,507
    2017 - 163,061.42 
    2018 - 62,146.9
Results: 
    2016 - 299,507.0
    2017 - 163,061.42
    2018 - 621,46.9 

Peformed  check on Spending Over Geography - POP/Recipient State, County, District

Test 2 - Filter on FY 2017 and Awarding Agencies (SBA and TREAS)
Expected: 2017 - 1,239,443.48
Result:  2017 - 1,239,443.48

Peformed  check on Spending Over Geography - POP/Recipient State, County, District

Test 3 - Award Amount Filter, 6k-85k, FY 2017
Expect Award Counts: Contracts 612, Grants 18, Direct Payments 72, Loans 6, Other 69
Result: Matched row counts in award download files
```
- [x] API Performance evaluation completed and present 
```
    Testing based on averages running on local server, dev rds instance
    
    Timing changes:
     1. Spending By Award
        Before = 1.08 seconds
        After =  1.18 seconds
    
     2. Spending over time
        Before = 1.05 seconds
        After = 1.07 seconds
    
     3. Spending By Geography
        Before = 1.08 seconds
        After = 1.12 seconds

     4. Spending By Category
        Before = 1.19 seconds
        After = 1.14 seconds

     5. Award amounts matview filter
        Before = 0.5 seconds
        After = 1.08 seconds

    Data volume changes:
        N/A
 ```